requirements-minimal.txt
# Минимальные зависимости для Telegram бота
python-telegram-bot==20.7
requests==2.31.0

# AI/ML APIs (выбираем одну)
groq==0.4.1

# Утилиты
python-dotenv==1.0.0

# Опциональные зависимости для расширенного функционала
# Раскомментируйте при необходимости:
# Flask==3.0.0  # для веб-интерфейса
# Pillow==10.1.0  # для обработки изображений
# numpy==1.24.3  # для семантического поиска
# scikit-learn==1.3.2  # для TF-IDF поиска